Fix comparison that is always false.
read_offset_code returns a signed integer but the result is stored in
an unsigned variable, so the returned error of -1 cannot be detected.
Change to a signed variable instead and cast appropriately (thanks
Thomas Klausner).

+++ lib/lh_new_decoder.c
@@ -489,7 +489,7 @@ static void output_byte(LHANewDecoder *decoder, uint8_t *buf,
 static void copy_from_history(LHANewDecoder *decoder, uint8_t *buf,
                               size_t *buf_len, size_t count)
 {
-	size_t offset;
+	int offset;
 	unsigned int i, start;
 
 	offset = read_offset_code(decoder);
@@ -498,7 +498,8 @@ static void copy_from_history(LHANewDecoder *decoder, uint8_t *buf,
 		return;
 	}
 
-	start = decoder->ringbuf_pos + RING_BUFFER_SIZE - offset - 1;
+	start = decoder->ringbuf_pos + RING_BUFFER_SIZE
+	      - (unsigned int) offset - 1;
 
 	for (i = 0; i < count; ++i) {
 		output_byte(decoder, buf, buf_len,
